requirements.txt
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
dnspython==2.4.2
docutils==0.20.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pycparser==2.21
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
urllib3==2.0.4
Werkzeug==2.3.7
